More schedule enhancements

+++ xSchedule/PlayList/PlayListItemPJLinkPanel.cpp
@@ -26,7 +26,7 @@ END_EVENT_TABLE()
 void PlayListItemPJLinkPanel::SetChoiceFromString(wxChoice* choice, std::string value)
 {
     choice->SetSelection(-1);
-    for (int i =0; i < choice->GetCount(); i++)
+    for (size_t i =0; i < choice->GetCount(); i++)
     {
         if (choice->GetString(i) == value)
         {
